Deduplicate ast expression walking logic
